requirements.txt
# Core AI/ML dependencies
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.18
faiss-cpu==1.7.4
transformers==4.35.0
torch==2.1.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3

# Data ingestion
requests==2.31.0
python-jira==3.5.0
datadog-api-client==2.23.0
sentry-sdk==1.38.0

# Test automation
appium-python-client==3.1.0
selenium==4.15.2
pytest==7.4.3
pytest-asyncio==0.21.1

# Orchestration & Infrastructure
kubernetes==28.1.0
temporalio==1.3.0
celery==5.3.4
redis==5.0.1

# Vision & Image Processing
opencv-python==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Development
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
